pyproject.toml
[tool.poetry]
name = "hi"
version = "0.1.0"
description = ""
authors = ["femto <femtowin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pyyaml = "^6.0.1"
[tool.poetry.scripts]
hi = 'hi:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
